Move broadcast and expand from arith to alias.
The latter seems to be a better home for what's there already.
